+++ Virtual_librarian/ReactAPP/ReactAPP.csproj
@@ -150,20 +150,24 @@
     <Content Include="Scripts\bootstrap.js" />
     <Content Include="Scripts\bootstrap.min.js" />
     <None Include="Scripts\jquery-1.10.2.intellisense.js" />
-    <Content Include="Scripts\react\Components\Archives.js" />
+    <Content Include="Scripts\react\Components\authenticationBackground.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Archives.js" />
     <Content Include="Scripts\react\Components\authentication\AuthenticationStyling.css" />
-    <Content Include="Scripts\react\Components\Contacts.js" />
-    <Content Include="Scripts\react\Components\Devices.js" />
-    <Content Include="Scripts\react\Components\Featured.js" />
-    <Content Include="Scripts\react\Components\Footer\Footer.js" />
-    <Content Include="Scripts\react\Components\Footer\_sticky-footer.css" />
-    <Content Include="Scripts\react\Components\Header\Title.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Contacts.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Devices.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Featured.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Footer\Footer.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\Footer\_sticky-footer.css" />
+    <Content Include="Scripts\react\Components\LogedInPage\Header\Title.js" />
     <Content Include="Scripts\react\Components\Header.js" />
     <Content Include="Scripts\react\Components\Home.js" />
     <Content Include="Scripts\react\Components\homestyle.css" />
-    <Content Include="Scripts\react\Components\images\sharon-mccutcheon-532782-unsplash.png" />
-    <Content Include="Scripts\react\Components\LibraryObjects\index.css" />
-    <Content Include="Scripts\react\Components\Layout.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\images\sharon-mccutcheon-532782-unsplash.png" />
+    <Content Include="Scripts\react\Components\LogedInPage\images\Virtual-Librarian-Homepage.png" />
+    <Content Include="Scripts\react\Components\LogedInPage\images\virtual-librarian-main-page.png" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryHome.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryObjects\index.css" />
+    <Content Include="Scripts\react\Components\LogedInPage\Layout.js" />
     <Content Include="Scripts\jquery-1.10.2.js" />
     <Content Include="Scripts\jquery-1.10.2.min.js" />
     <None Include="Scripts\jquery.validate-vsdoc.js" />
@@ -173,14 +177,14 @@
     <Content Include="Scripts\jquery.validate.unobtrusive.min.js" />
     <Content Include="Scripts\modernizr-2.6.2.js" />
     <Content Include="Scripts\react\app.js" />
-    <Content Include="Scripts\react\Components\LibraryObjects\Library.js" />
-    <Content Include="Scripts\react\Components\LibraryObjects\LibraryBooks.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryObjects\Library.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryObjects\LibraryBooks.js" />
     <Content Include="Scripts\react\Components\authentication\Register.js" />
-    <Content Include="Scripts\react\Components\LibraryObjects\LibraryStyling.css" />
-    <Content Include="Scripts\react\Components\margins.css" />
-    <Content Include="Scripts\react\Components\Settings.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryObjects\LibraryStyling.css" />
+    <Content Include="Scripts\react\Components\LogedInPage\margins.css" />
+    <Content Include="Scripts\react\Components\LogedInPage\Settings.js" />
     <Content Include="Scripts\react\Components\authentication\SignIn.js" />
-    <Content Include="Scripts\react\Components\LibraryObjects\Utils.js" />
+    <Content Include="Scripts\react\Components\LogedInPage\LibraryObjects\Utils.js" />
     <Content Include="Scripts\react\dist\bundle.js" />
     <Content Include="Scripts\react\eslintrc.js" />
     <Content Include="Scripts\react\webpack.config.js" />
