--- Virtual_librarian/ReactAPP/Scripts/react/Components/Archives.js
@@ -1,14 +1,0 @@
-﻿import React from "react"
-
-export default class Archives extends React.Component {
-
-    render() {
-        
-        return (
-           
-                <h1>Kazkas</h1>
-                
-        );
-    }
-}
-Archives.id = "app";
