--- Virtual_librarian/ReactAPP/Scripts/react/Components/Contacts.js
@@ -1,19 +1,0 @@
-﻿import React from "react";
-import "./margins.css"
-import {
-    Container, Col, Form,
-    FormGroup, Label, Input,
-    Button,
-} from 'reactstrap';
-
-export default class Register extends React.Component {
-
-    render() {
-
-        return (
-            <div className="outside">
-                <h1>Comming soon</h1>
-                </div>
-        );
-    }
-}
