--- Virtual_librarian/ReactAPP/Scripts/react/Components/Featured.js
@@ -1,12 +1,0 @@
-﻿import React from "react"
-import "./margins.css"
-export default class Featured extends React.Component {
-
-    render() {
-        return (
-            <div>
-                <h1 >Featured</h1>
-                </div>
-        );
-    }
-}
