--- Virtual_librarian/ReactAPP/Scripts/react/Components/Footer/_sticky-footer.css
@@ -1,46 +1,0 @@
-﻿.Site {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-    background-color: black;
-    font-size:60pt;
-}
-
-.Site-content {
-    flex: 1 0 auto;
-    padding: var(--space) var(--space) 0;
-    width: 100%;
-    background-color: black;
-}
-
-    .Site-content:after {
-        width: 66%;
-        margin: 500px;
-        font-size: 100px;
-        position: absolute;
-        left: 0;
-        bottom: 0;
-        right: 0;
-    }
-.center {
-    margin: auto;
-    background: none;
-    width: 66%;
-}
-
-.outside {
-    margin: 100px;
-    background-color: black;
-    
-}
-.backgroundCol {
-    background-color: #1f257c;
-    height: 100px;
-    text-align: center;
-    font-size:15px;
-    color: aliceblue;
-    position: fixed;
-    left: 0;
-    bottom: 0;
-    right: 0;
-}
