+++ Virtual_librarian/ReactAPP/Scripts/react/Components/Header.js
@@ -1,5 +1,5 @@
 ﻿import React from "react";
-import Title from "./Header/Title";
+import Title from "./Header/LogedInPage/Title";
 
 export default class Header extends React.Component {
