--- Virtual_librarian/ReactAPP/Scripts/react/Components/LibraryObjects/LibraryBooks.js
@@ -1,71 +1,0 @@
-﻿import React from "react";
-import { render } from "react-dom";
-import { makeData, Logo, Tips } from "./Utils";
-
-// Import React Table
-import ReactTable from "react-table";
-import "react-table/react-table.css";
-import "./LibraryStyling.css"
-
-
-export default class LibraryBooks extends React.Component {
-    constructor() {
-        super();
-        this.state = {
-            data: makeData()
-        };
-    }
-    render() {
-        const { data } = this.state;
-        return (
-            <div className="outside">
-                <ReactTable
-                    data={data}
-                    columns={[
-                        {
-                            Header: "Name",
-                            columns: [
-                                {
-                                    Header: "First Name",
-                                    accessor: "firstName"
-                                },
-                                {
-                                    Header: "Last Name",
-                                    id: "lastName",
-                                    accessor: d => d.lastName
-                                }
-                            ]
-                        },
-                        {
-                            Header: "Info",
-                            columns: [
-                                {
-                                    Header: "Age",
-                                    accessor: "age"
-                                },
-                                {
-                                    Header: "Status",
-                                    accessor: "status"
-                                }
-                            ]
-                        },
-                        {
-                            Header: 'Stats',
-                            columns: [
-                                {
-                                    Header: "Visits",
-                                    accessor: "visits"
-                                }
-                            ]
-                        }
-                    ]}
-                    defaultPageSize={10}
-                    className="-striped -highlight"
-                />
-                <br />
-                <Tips />
-                <Logo />
-            </div>
-        );
-    }
-}
