--- Virtual_librarian/ReactAPP/Scripts/react/Components/LibraryObjects/LibraryStyling.css
@@ -1,15 +1,0 @@
-﻿.center {
-    background: none;
-    width: 66%;
-    padding: var(--space) var(--space) 0;
-    margin: 100px;
-}
-
-.outside {
-    margin: 100px;
-    padding: var(--space) var(--space) 0;
-    background: none;
-    width: 66%;
-}
-
-
