+++ Virtual_librarian/ReactAPP/Scripts/react/Components/LogedInPage/Archives.js
@@ -1,0 +1,14 @@
+﻿import React from "react"
+
+export default class Archives extends React.Component {
+
+    render() {
+        
+        return (
+           
+                <h1>Kazkas</h1>
+                
+        );
+    }
+}
+Archives.id = "app";
