+++ Virtual_librarian/ReactAPP/Scripts/react/Components/LogedInPage/Header/Title.js
@@ -1,0 +1,11 @@
+﻿import React from "react"
+
+export default class Title extends React.Component {
+
+    render() {
+        return (
+
+            <h1>{this.props.title}</h1>
+        );
+    }
+}
