+++ Virtual_librarian/ReactAPP/Scripts/react/Components/LogedInPage/LibraryObjects/index.css
@@ -1,0 +1,7 @@
+﻿body {
+    margin: 100px;
+    padding: var(--space) var(--space) 0;
+    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
+    font-weight: 300;
+    font-size: 14px;
+}
