--- Virtual_librarian/ReactAPP/Scripts/react/Components/Settings.js
@@ -1,16 +1,0 @@
-﻿import React from "react"
-import "./margins.css"
-export default class Settings extends React.Component {
-
-    render() {
-        return (
-            
-            <div className="outside">
-                <h1>Settings</h1>
-                </div>
-
-                
-        );
-    }
-}
-Settings.id = "app";
