+++ Virtual_librarian/ReactAPP/Scripts/react/Components/authentication/AuthenticationStyling.css
@@ -5,7 +5,7 @@
 }
 .outside {
     margin: 5rem 10rem 10rem 25rem;
-    background: none;
+  
     width: 66%;
 }
 .buttonPlace{
