+++ Virtual_librarian/ReactAPP/Scripts/react/Components/authentication/SignIn.js
@@ -1,10 +1,10 @@
 ﻿
 import React, { Component } from "react";
-import "./AuthenticationStyling.css"
+import "../LogedInPage/margins.css"
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 //import { Button, Radio, Icon } from 'antd';
-
+import { Link } from "react-router-dom";
 import 'antd/dist/antd.css';
 import { withStyles } from '@material-ui/core/styles';
 import MenuItem from '@material-ui/core/MenuItem';
@@ -12,10 +12,10 @@ import TextField from '@material-ui/core/TextField';
 import {
     Container, Col, Form,
     FormGroup, Label, Input,
-    Button
+    Button,
 } from 'reactstrap';
 /*Text box*/
-const styles = theme => ({
+/*const styles = theme => ({
     container: {
         display: 'flex',
         flexWrap: 'wrap',
@@ -23,6 +23,7 @@ const styles = theme => ({
     textField: {
         marginLeft: theme.spacing.unit,
         marginRight: theme.spacing.unit,
+        
         width: 200,
         fontSize: 50,
     },
@@ -35,7 +36,7 @@ const styles = theme => ({
     resizeFont: {
        fontSize: 50,
     },
-});
+});*/
 
 export default class SignIn extends React.Component{
     /*Button*/
@@ -57,40 +58,50 @@ export default class SignIn extends React.Component{
     
     render() {
         const size = this.state.size;
-        const classes = styles;
+        //const classes = styles;
         return (
-            <div className="outside">
-               
-                <Container>
+           
+           
+
+            <Container>
+                <div className="homepage">
                     <h2>Sign In</h2>
                     <form noValidate autoComplete="off">
                         <TextField
                             id="Username"
                             label="Username"
                           
-                            className={classNames(classes.textField, classes.dense)}
+                            
                             margin="dense"
-                            style={{ marginLeft: 8, marginRight: 500, marginTop: 8, marginBottom: 8, width: 300, fontSize: '<50>', }}
+                        style={{ marginRight:-120,marginTop:20, fontSize: '<50>', }}
                         />
                     <TextField
                             id="standard-password-input"
-                            className={classes.textField}
-                            style={{ marginLeft: 8, marginRight: 500, marginTop: 8, marginBottom: 8, width: 300, fontSize: '80em', }}
+                            
+                            style={{ marginTop:80,  fontSize: '80em', }}
                         label="Password"
                         type="password"
                         title={this.state.name}
                             
                 margin="normal"
-                autoComplete="current-password"
-            />
+                            autoComplete="current-password"
+
+                        />
+                        
                        
                         
                     </form>
-                    <Button color="primary">Login</Button>
+                    <Link to="/library/home">
+                        <Button color="primary">SignIn</Button>
+                        </Link>
+                    <Link to="/register">
+                        <Button color="primary">Register</Button>
+                    </Link>
+                </div>
                 </Container>
               
                
-            </div>
+           
         );
     }
 }
